Revert "Fix strict-alias breaking warnings with GCC."
This reverts commit 0d3044b547614cbd313d90021606af1f81fb10de.

I'm not sure if this is good for anything. I can't reproduce the
failures that happen on Mac and Windows, so let's try reverting the
only patch that touches QtDeclarative.

If this works, then we'll have found out that the code is broken and
my patch only revealed the errors.