Fix missing color in text when using static text back-end in QML
Using QStaticTextItem as a back-end for QML text on GL caused a
regression where parts of a text element would get the wrong color.
This was because the color set on the painter which was passed into
draw() was never transferred to the painter used to record the draw
text calls issued by the underlying QTextLayout::draw()-function.

Task-number: QTBUG-18428
Reviewed-by: Jiang Jiang