Fix qfile autotest hanging in CI system
The test machine is configured to pop up dialogs in case of CRT
assertion failures. The CRT has a debug assert whenever an invalid
file handle is used, while the test is expecting to get the EBADF
return code.
Due to some behaviour change in windows 7 (maybe SP1), we need to
call _CrtSetReportMode in addition to setting an error handler.

Reviewed-By: mread