Drift correction and better accuracy for repeating timers in Symbian
Timers on Symbian were always firing at least 2ms late. This was partly
due to a paranoid extra delay in Qt and the kernel being very cautious.

For one shot timers this is not so bad although a bit too much. But for
repeating timers, this can cause significant reductions in firing rate
particularly for short period timers.

The timer active object now corrects timer lateness by up to 4ms per
event. This is enough to compensate for the kernel lateness without
the possibility of wild corrections.

Task-number: QTBUG-18549
Reviewed-by: iain