Build break fix for simulated QS60Style
Recent changes in QS60Style (to support placeholder background texture)
causes the simulated style not to build. The implementation
for new method placeHolderTexture() is on the Symbian-specific
sourcefile, which is omitted in the simulator builds.

As a fix, move the implementation to the "generic" style source
file qs60style.cpp, since the method does not contain any
Symbian specific code.

Task-number: QTBUG-18863
Reviewed-by: owolff