QDeclarativeObjectScriptClass: do not handle valueOf()
The obj.valueOf() is called in situations that we convert an object to
a number. It is called implicitly by V8 (see DefaultNumber() in
runtime.js for example), and when it hits our ScriptClass code the
frame doesn't look like the way we expect (no DeclarativeContext
available in the Caller scope chain).

The problem happen in callQtInvokables, in tests that expect
numbers and we pass the an object instead.

This was the only case I've found which v8::Context::GetCallerContext
(an extension in our branch of v8) isn't enough for the purpose of
providing the top context in the previous frame. Avoiding valueOf() is
OK for now.

Last but not least, tst_qdeclarativeecmascript does not segfault anymore.

Reviewed-by: Olivier Goffart