Fix memory leak in XSD component of XmlPatterns
There are couple of memory leaks in XmlPatterns when parsing XSD
files. The reason is that the module over uses the shared data
pointer and generates many cyclic reference loop. Some part of
the code is refactored. Instead of using shared data pointer,
normal C++ pointer is used to break the ownership loop. This is
bug fix for QTBUG-8948.

Task-number: QTBUG-8948
Reviewed-by: Sami Merila and Laszlo Agocs