Made the autotest for drawing pixmaps with painter open more fail safe.
The autotest relies on QPixmap::grabWindow() which causes random CI
failures because it may capture bogus content if some other window
comes to foreground, focus is changed, etc. To overcome these false
positives the pixmap content comparison is now only done when the
captured content is one of the three possible images. If it is
anything else, we cannot verify so the case is skipped.

Task-number: QT-4002
Reviewed-by: TRUSTME