Fix missing glyphs for large fonts with QStaticText/GL/Freetype
When rendering fonts that are so large the Freetype engine will fall
back to outline rendering, we would sometimes lose glyphs that were
at the edge of the texture, because the width of the image returned
by QFontEngine::alphaMapForGlyph() includes a margin (which the other
font engines do not) and thus it would sometimes be wider than the
available area in the cache's texture. This would in turn cause the
GL call to copy the glyph into the cache to fail.

The correct fix would probably be to make the algorithm used to
calculate the width in QFontEngine::alphaMapForGlyph() the same as
the algorithm used by the glyph cache (which would mean to remove
the margin there), but to minimize the impact in a patch release,
we fix the symptom and move the correct fix to next minor release
instead.

Task-number: QT-4876
Reviewed-by: Jørgen