Remove Q_ASSERT's from QObject autotest.
The Receiver class has two slots that aren't meant to get called during
the test (they're there to catch broken parsing of slot names).  Rather
than asserting when one of them gets called, which does nothing in a
release mode build, this commit makes the slots record the number of
times they were called (as for the other slots in the test) and fails
the test gracefully if either of those slots was called.

Change-Id: Ia0393026cb96ffdc6190b5e7bd951f75d231b11e
Task-number: QTBUG-17582
Reviewed-by: Rohan McGovern