Remove Q_ASSERT's from QScriptable autotest
The helper class expects a valid script engine to be available and
attempted to verify this with Q_ASSERT (which does nothing in release
mode) and QVERIFY (which should only be used directly in test
functions).  This commit makes the helper class avoid using an invalid
engine and catches the problem later by verifying that the last script
engine used by the scriptable class was valid.

Change-Id: I849df32b141b7801b9b5a92f44ff325397e32f2d
Task-number: QTBUG-17582
Reviewed-by: Rohan McGovern