Remove Q_ASSERT's from modeltest
Rather than aborting in debug builds and ignoring the failures in
release builds, report specific warnings on each failure, count the
failures, and fail the test if the failure counts are non-zero at the
end of the test.

The same change is also made for a QCOMPARE that appeared
inappropriately inside a helper class.  QCOMPARE may only appear
directly in a test function.

Change-Id: I81f0ce80512fa72c67f5aa72c0511a4b650d5d20
Task-number: QTBUG-17582
Reviewed-by: Rohan McGovern