Disable QtConcurrent for Symbian winscw builds.
While QtConcurrent is nowadays usable with RVCT and GCCE compilers,
the compiler used for the winscw target has problems with the
templates still. To enable building 4.8 on winscw, the patch
defines QT_NO_CONCURRENT and QT_NO_QFUTURE in qglobal.h for
this target. However this causes further complications in the
Q_OBJECT classes of QtConcurrent because moc will be run without
having QT_NO_CONCURRENT and QFUTURE defined (as moc is not a Symbian
app and for other Symbian targets they must not be defined, which makes
differentiation in qglobal.h impossible for the moc run) and having a
moc file generated for a class that will be ifdef'ed out during
compilation causes build breaks. Therefore additional dummy stubs are
provided in QFutureWatcherBase.

Reviewed-by: Liang Qi