QDeclarativeMouseArea: block context menu events
If the MouseArea accepts the same button used to trigger a context
menu event, it should not let the event to be delivered to item
behind.

This is important for items that do implement contextMenuEvent(), like
QDeclarativeWebView. When there's a mouse area on top of this item and
that accepts the right click (in Linux), the event was still being
delivered and the WebView menu was incorrectly appearing.

QtWebKit bug https://bugs.webkit.org/show_bug.cgi?id=56526 documents this
problem.

Change-Id: I386fac6c96f47b8616e2eeb7e5f97043ba418980
Merge-request: 1185
Reviewed-by: Martin Jones <martin.jones@nokia.com>