X11: Keep the menubar inside the widgetbox window in toplevel mode
For now the appmenu protocol does not make it possible to associate a
menubar with all application windows. This means in top level mode you
can only reach the menubar when the widgetbox window is active. Since
this is quite inconvenient, better not use the native menubar in this
configuration and keep the menubar in the widgetbox window.

Merge-request: 1229
Reviewed-by: denis