Support word selection list with predictive text from splitview
When splitview is active and user taps a predicted word, mouse button
should be forwarded to m_pointerHandler, which opens a suggested word
list. When splitview is not active, but there are HW QWERTY keys in the
device, tapping a word, should only move the cursor. Without HW QWERTY
and no splitview, native editing state handles the functionality.

Task-number: QTBUG-19062
Reviewed-by: Miikka Heikkinen