Fixed crashes in Observer mode related to infinite bounding rects
Bug fixed by avoiding uniting or subtracting QPolygonF with potentially
infinite coordinates.

The LiveSelectionIndicator now uses a QGraphicsRectItem rather than a
QGraphicsPolygonItem and displays only the boundaries of selected
objects, not including their children.

The SubcomponentMaskLayerItem now works with rectangles and uses a
QRegion to determine the area around the current context, converting
this to a polygon only as a last step.

Reviewed-by: Kai Koehne
Task-number: QTCREATORBUG-4559
Change-Id: I266f5387fa67017fc50215282a95b4ee6498be6d