Symbian socket: Implement systemProxyForQuery for service networks
When the specified (or active/default) network is a service network,
try to determine the appropriate IAP network within it.
This is the highest priority active network (if any)
followed by the highest priority discovered network (if any)
or the highest priority network as a last resort.

As Symbian doesn't include proxy type in the comms database, proxies
are assumed to be http type unless the port is 1080 (iana assigned
port for SOCKS) in which case the proxy is set as SOCKS type.

Also fixed unitialised member issue with SymbianIapId internal class

Task-number: QTBUG-18618
Reviewed-by: Markus Goetz