tests: allow unstable tests to be marked with CONFIG+=insignificant_test
Marking a test with CONFIG+=insignificant_test will cause the exit code
of the test to be discarded during `make check'.  This is intended to be
used for tests which are valuable to run, but are known to be unstable
and are not feasible to immediately fix.

Reviewed-by: Jason McDonald
Change-Id: I50a712c33c2ebb0af39f1ea0bf2adef7f0936425
(cherry picked from commit b0a5a988900b777cb08012408eb108512a3e873b)