Fix in-process drag-n-drop of image data, image/* was not available.
If we give the exact initial QMimeData to the dropEvent, we only get
application/x-qt-image as available mimeType. We need to go through QDropData
to call xdndObtainData, which can still do some in-process optimization,
but there we can do the "saving QImage to the requested format" conversion.

Task-number: QTBUG-4110

Merge-request: 860
Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>