Missing glyphs transforming QStaticText on X11/raster with subpixel AA
Static text took an untested and broken code path for the combo of
subpixel AA, X11, raster engine and transformation. This would cause
missing glyphs. The reason was that QStaticText took an unused code path
which turned out not to work. The workaround is to use gray AA on
transformed text, like we already do for the GL engine. In Qt 4.8, the
static text code path has been rewritten to use the Freetype cache
instead of the image glyph cache, so the bug will be fixed more properly
there.

Reviewed-by: Samuel