Fix glyph metrics with QStaticText/Freetype/raster and light/no hinting
This is a back-port of part of cad70d64d0bbada. In the raster engine's
drawCachedGlyphs(), which is used by QStaticText, we would use the wrong
metrics to lay out the glyphs, because loadGlyphMetrics() would assume
full hinting. A visible effect of this was that the baseline of rotated
text became wavy.

Task-number: QTBUG-18185
Reviewed-by: Jiang Jiang