Fixed compile of tst_qscriptextensionplugin on some Windows configurations
The debug and release versions of staticplugin are qmake'd with the
same destination directory.  This causes the generated staticplugin .prl
file to always refer to the debug versions of Qt libraries, even if Qt
was configured with -release.

The .prl mechanism is not useful for this test, so simply disable it to
solve the problem.

Reviewed-by: ckamm
(cherry picked from commit b761f2e1b323012661a9aeed3873b1e5facb0f11)