Fixed compile of tst_qscriptextensionplugin on some Windows configurations
The debug and release versions of staticplugin are qmake'd with the
same destination directory.  This causes the generated staticplugin .prl
file to always refer to the debug versions of Qt libraries, even if Qt
was configured with -release.

The .prl mechanism is not useful for this test, so simply disable it to
solve the problem.

Reviewed-by: ckamm
(cherry picked from commit f5a63feb8953799de7e787f333575ee37fae8a3f)