tst_qeventloop runs the event loop to ensure socket messages arrive
Auto test tst_qeventloop assumed that socket messages would be received
within a single call to processEvents(). But this is not necessarily
the case, and in particular with the combination of Symbian round robin
active scheduler and new socket engine. So tst_qeventloop now uses an
event loop to complete the test rather than just a single call to
processEvents().

Task-number: QTBUG-15019
Reviewed-by: brad