symbian socket engine: resolve some fixme and todo items
Should be no behaviour changes from this commit, one function is renamed
and otherwise it is comment changes. (stuff we were unsure about when
creating the code initially but have decided to keep)

Multicast is not fully supported by symbian, and out of scope for the
QTBUG-7274 task. So the two stub functions are left with the ### comment
for the benefit of a future implementor.

Task-number: QTBUG-18371
Reviewed-by: Markus Goetz