Build fix for Mac OS 10.5
The build breaks because QString defines a template function
that interferes with a system header. The easy fix is to just
make sure we include the system headers before any Qt headers.
This fix turned out to already be in for Qt5, but that change
contained other stuff as well, and were not suited for
cherry-picking.

Rev-By: msorvig