Symbian: Fix QFontInfo::pixelSize()
Unlike QFont::pixelSize(), which may return -1 if the font size was
defined in points, QFontInfo::pixelSize() always needs to return a
valid value.

c4ef479906f073fa84999eb950f00e264ebd4e8e which was a fix for
QTBUG-13009 tried to fix a similar issue, but failed to do that
properly, which resulted in QTBUG-15513 and QTBUG-17844.

This commit is supposed to fix all three bugs.

Task-Number: QTBUG-13009
Task-Number: QTBUG-15513
Task-Number: QTBUG-17844