Mac: switch raster off as default paint engine
After discussing the state of the raster engine with many key devs in
Oslo, we have concluded the following:

Switching to raster as default is a big change for Qt-4.8. The
posibilty of introducing regression for 3rd-party applications
are high. From manual testing we can easily spot regressions
ourselves, expecially for text rendering. And the overall drawing
performance seems to drop slightly. So there seem to be no reason
for us to take such a risk at this point in time for Qt-4.8, expecially
since we have other tasks that should get our attention going
forward, and the main person that did the implementation has left.

Rev-By: Gunnar Sletta
Rev-By: Fabien Freling