QTBUG-19883 Adding top level TRAP for QThreads on Symbian
The native Symbian implementation of QThread in Qt4.8 does not have
a top level TRAP, whereas it did in pthreads based implementation in
Qt4.7. This causes an incompatibility in the form of a EUSER-CBASE:66
panic for code that attempts to use the cleanup stack without a TRAP
of its own.

This adds a top level TRAP and std::exception handler to match it and
prints out warning information if they ever trigger to provide some
diagnostic information which would otherwise be lost.

Task-number: QTBUG-19883
Reviewed-by: Laszlo Agocs