Resizable graphicsview's background is drawn incorrectly in splitview
Due to the fact that native side re-opens keyboard before sending
rotation event, the original size of graphicsview before auto-translate,
is stored incorrectly. We store the height of graphicsview in previous
orientation. Now, if the window was maximized, graphicsview was drawn
correctly, since closing vkb would change the window state back to
maximized and thus force a resizing of the view. With fullscreen gv
this didn't happen. Fullscreen graphicsview thus needs a forced
resize back to fullscreen.

Task-number: QTBUG-19856
Reviewed-by: Miikka Heikkinen