Revert "Symbian: Fix QFontInfo::pixelSize()"
This reverts commit fcfc19878a0a1a48194a786bba64da11606077d2.

I am happy that this commit fixed three bugs at once. But
Actually, I am not sure if QTBUG-15513 should be fixed at
this point. Fact is that the patch as it is would have changed
the point->pixels calculation back to how it was in Qt 4.6.

This means that the fonts which are defined with pointSize
would now (in Qt 4.7.4) suddenly be bigger than they were
in Qt 4.7.3. Imho this is unacceptable, as it would break
all layouts which were developed for Qt 4.7 apps, when
point size (instead of pixle size) was used.

I will need to fix QTBUG-17844 without fixing QTBUG-13009

If QTBUG-13009 will be fixed for 4.8 will be discussed.