Workaround webkit deadlock on macos x
The webkit AtomicallyInitializedStatic and Qt's Q_GLOBAL_STATIC can
deadlock on the Mac, as the mac compiler inserts calls to
__cxa_guard_acquire and __cxa_guard_release around initialisation of
local statics.
In Q_GLOBAL_STATIC case, this is the QGlobalStaticDeleter local static
Whereas webkit AtomicallyInitializedStatic is a local static variable
in any case.

Problem is triggered because webkit constructs QNetworkConfigurationManager
inside the constructor of a local static - networkStateNotifier
And the generic bearer plugin calls QNetworkInterface::allInterfaces
in the bearer thread, which needs an initialised Q_GLOBAL_STATIC.

Reviewed-by: Laszlo Agocs