QTBUG-17776, reporting terminated threads as not running on Symbian
On Symbian app shutdown all threads are terminated and their stack
memory is released, but there is no time for notification of exit of
these threads. So any attempt to access stack data in such a thread
from static data destruction will cause a crash.

This was happening with the XmlQuery thread. It was testing if the
thread was still running, and QThread thought it was because there was
no notification. So when the XmlQuery thread was asked to exit, and
QThread::exit tried to access a stack based QEventLoop, there was a
crash.

By adding a test if the thread has been terminated to
QThread::isRunning(), clients can now rely on this to know that it is
safe to call exit() on a thread. The existing code is made safe again.

Task-number: QTBUG-17776
Reviewed-by: Shane Kearns