Splitview - Auto-translation rules changed
When using "splitview" (virtual keyboard with non-fullscreen
editing mode), it is currently auto-translating the cursor to the
center of the screen if possible.

It would be preferable, if the translation would only be minimal,
just enough for cursor to be visible. This makes scrolling of
input widget (i.e. large editor) easier to use, as text flows
naturally (row-by-row) and not in "jumps" like it used to do.

Additionally, limit the translation to the end of input widget
boundary.

Task-number: QTBUG-20034
Reviewed-by: Miikka Heikkinen