qmlplugindump: Fix dumping with -path on Mac.
If the current working directory was a direct parent of the qmldir path
the exported modules had the path as the module URI on macs.

Also changes the QtQuick export back to 1.0 to make it work with
Qt 4.7.3. The version of that import statement does not actually matter
as long as it's valid.

Mirrors a change in Qt Creator:
http://codereview.qt.nokia.com/896

Reviewed-by: Kai Koehne