Cocoa: fix qwidget auto test failures
Some of the test are bound to fail then using the native paint
engine on Mac. They failed before as well, but were swiched on
as a part of the raster engine implementation. But now that we
decide not to go with raster as default after all, be skip some
of the tests again

Rev-By: msorvig