Cocoa: fix qtabwidget auto test failure
The fixed test fails when switching back to native paint engine
as default, as opposed to raster. I'm not really sure why, and I
hate just commenting the test out. But right now it is more important
to 'revert' raster as default for Qt-4.8-beta, than to puzzle with
update optimizations for tab widget.

Rev-by: msorvig