Fix headers generation for QtWebKit.
syncqt assumes that the pro files of modules are named the same way as
the directory they are in. QtWebKit 2.2 has its main pro file called
QtWebKit.pro and not qt.pro even if it's located in src/3rdParty/webkit/Source/WebKit/qt.
This patch ensure that syncqt will look for a pro file if there is one in
the directory in the case there isn't one named the same way as the directory.

Reviewed-by:mariusSO