Symbian socket engine: remove remaining todo comments
The writes >16k blocking in the emulator only applies to the winsock
connectivity used in S60 SDKs. It doesn't affect the ethernet
connectivity used by platform environments.

Restarting notifier after error seems like the correct thing to do,
and isn't causing any problems.

The duplicated code for setting error strings is unfortunate, but a
consequence of our decision not to derive from the native socket engine.
If symbian ever gets a Qt5 port, we should revisit it there.

Task-number: QTBUG-18371
Reviewed-by: Markus Goetz