Enable key capture and RemCon interfaces for Qt apps on Symbian
The volume and other multimedia keys in Symbian are not delivered
through the normal key events, but through a seperate API called
CRemConCoreApiTarget. The commit implements the feature that multimedia
key events are delivered via normal key events to Qt Application, if
the Qt::AA_CaptureMultimediaKeys is defined.

Task-number: QTBUG-4415
Reviewed-by: Sami Merila
Reviewed-by: Miikka Heikkinen