Flicking behaviour of ListView/GridView SnapOnItem is inconsistent
Improve the response of the views when SnapOneItem/Row is enabled.
In this case it is best to be much more reactive to the user input
since even a small movement in a particular direction indicates
a change to the next/previous item.

Change-Id: I6a8eb689c3b12cdc67f24106032e36bba82d2846
Task-number: QTBUG-19874
Reviewed-by: Bea Lam
(cherry picked from commit e7bebcf0c59368340df524db4a53ae2595d057d7)